fix(hardware): cast ulong Sum to long + pass ComputerInfo to RenameViewModel
Two follow-on issues from the build failure:
- HardwareReportViewModel.BuildMemorySummary: Enumerable.Sum has no
  ulong overload, so the lambda must cast to long.
- MainViewModel: RenameViewModel now takes a ComputerInfo (not a string
  computer name) so callers can update every field after a rename.

+++ src/ComputerRenameTool/ViewModels/HardwareReportViewModel.cs
@@ -92,7 +92,7 @@ public sealed class HardwareReportViewModel : ObservableObject
             return "数据不可读";
         }
 
-        var total = chips.Sum(c => c.CapacityBytes ?? 0UL);
+        var total = chips.Sum(c => (long)(c.CapacityBytes ?? 0UL));
         var totalGb = Math.Round(total / (1024d * 1024d * 1024d), 1);
         var modules = string.Join(" + ", chips.Select(c =>
         {
